whisper/tests/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
grpcio==1.60.0
